Change Startup into extensions
