+++ Falcon.BackEnd/Falcon.BackEnd.Security/Program.cs
@@ -1,6 +1,10 @@
 using Falcon.BackEnd.Security.Domain;
 using Falcon.Libraries.Microservice.Startups;
 
-var app = new Startup<ApplicationDbContext>(args);
+var builder = WebApplication.CreateBuilder();
 
-app.Run();
+builder.UseMicroservice<ApplicationDbContext>();
+
+var app = builder.Build();
+
+app.RunMicroservice<ApplicationDbContext>();
