+++ Falcon.BackEnd/Falcon.BackEnd.Showcase/Falcon.BackEnd.Showcases.csproj
@@ -13,7 +13,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Falcon.Libraries\Falcon.Libraries.Microservice\Falcon.Libraries.Microservice.csproj" />
     <ProjectReference Include="..\..\Falcon.Models\Falcon.Models.csproj" />
-    <ProjectReference Include="..\Falcon.BackEnd.Products\Falcon.BackEnd.Products.csproj" />
   </ItemGroup>
 
 </Project>
